+++ src/Calliope/Concerns/CallsApi.ts
@@ -7,8 +7,11 @@ import type Model from '../Model';
 import type { QueryParams } from './BuildsQuery';
 import BuildsQuery from './BuildsQuery';
 import type { Attributes, SimpleAttributes } from './HasAttributes';
-import { isObjectLiteral, transformKeys } from '../../Support/function';
-import { finish, kebab, plural } from '../../Support/string';
+import isObjectLiteral from '../../Support/function/isObjectLiteral';
+import transformKeys from '../../Support/function/transformKeys';
+import finish from '../../Support/string/finish';
+import kebab from '../../Support/string/kebab';
+import plural from '../../Support/string/plural';
 import type { MaybeArray, StaticToThis } from '../../Support/type';
 
 /**
